Break up libcarma
libcarma has been broken up into lc, sampler and carma.
