+++ python/k2.py
@@ -11,14 +11,16 @@ import argparse
 import matplotlib.pyplot as plt
 import pdb
 
+plt.ion()
+
 try:
-    import libcarma as libcarma
+    import lc
 except ImportError:
-    print 'libcarma is not setup. Setup libcarma by sourcing bin/setup.sh'
+    print 'kali is not setup. Setup kali by sourcing bin/setup.sh'
     sys.exit(1)
 
 
-class k2LC(libcarma.basicLC):
+class k2LC(lc.basicLC):
 
     sap = ['sap', 'raw', 'uncal', 'un-cal', 'uncalibrated', 'un-calibrated']
     pdcsap = ['pdcsap', 'mast', 'cal', 'calib', 'calibrated']
