+++ tests/test_computeLogLikelihood.py
@@ -4,7 +4,8 @@ import unittest
 import pdb
 
 try:
-    import libcarma
+    import lc
+    import carma
 except ImportError:
     print 'libcarma is not setup. Setup libcarma by sourcing bin/setup.sh'
     sys.exit(1)
@@ -15,12 +16,12 @@ class TestComputeLnLikeCARMA10(unittest.TestCase):
     def setUp(self):
         self.p = 1
         self.q = 0
-        self.nt = libcarma.basicTask(self.p, self.q)
+        self.nt = carma.CARMATask(self.p, self.q)
         self.dt = 1.0
         self.TAR1 = 5.0
         self.Amp = 1.0
         self.rho = np.array([-1.0/self.TAR1, self.Amp])
-        self.theta = libcarma.coeffs(self.p, self.q, self.rho)
+        self.theta = carma.coeffs(self.p, self.q, self.rho)
         self.nt.set(self.dt, self.theta)
 
     def tearDown(self):
